Move error utils to utils folder
