+++ packages/core/src/utils/errorUtils.ts
